requirements.txt
# Django e dependências principais
Django==4.2.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.2.0

# Banco de dados (somente PostgreSQL)
psycopg2-binary==2.9.10
dj-database-url==2.1.0

# Variáveis de ambiente
python-dotenv==1.0.0
python-decouple==3.8

# Celery para tarefas assíncronas
celery==5.3.4
django-celery-results==2.5.1
flower==2.0.1

# Utilitários
Pillow==10.0.0
django-filter==23.3

# Desenvolvimento
django-debug-toolbar==4.2.0
django-extensions==3.2.3

# Produção
gunicorn==21.2.0
whitenoise==6.5.0

# Testes
pytest==7.4.0
pytest-django==4.5.2
factory-boy==3.3.0

# Documentação da API
drf-spectacular==0.26.5

# Segurança
django-ratelimit==4.1.0
django-csp==3.7

# Cache
redis==4.6.0
django-redis==5.4.0

# Monitoramento
sentry-sdk==1.32.0
